requirements.txt
openai==1.78.1
langfuse==1.6.3
python-dotenv